+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Clipping issues with the unknown mods banner
 - Mod card lists such as the dependencies and hidden mods dialogs not being sorted correctly
 - Icons and links in update dialog always pointing to Thunderstore
+- Install button being stuck in a disabled state after installing mods
 
 ## 1.19.2 (2026-07-31)